Create SubscriptionsApi
