+++ src/NewTwitchApi/Resources/SubscriptionsApi.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace NewTwitchApi\Resources;
+
+use GuzzleHttp\Exception\GuzzleException;
+use Psr\Http\Message\ResponseInterface;
+
+class SubscriptionsApi extends AbstractResource
+{
+    /**
+     * @throws GuzzleException
+     * @link https://dev.twitch.tv/docs/api/reference/#get-broadcaster-s-subscribers
+     */
+    public function getBroadcasterSubscriptions(string $broadcasterId, array $ids = [], string $bearer): ResponseInterface
+    {
+        $queryParamsMap = [];
+
+        $queryParamsMap[] = ['key' => 'broadcaster_id', 'value' => $broadcasterId];
+
+        foreach ($ids as $id) {
+            $queryParamsMap[] = ['key' => 'user_id', 'value' => $id];
+        }
+
+        return $this->callApi('subscriptions', $queryParamsMap, $bearer);
+    }
+}
