+++ src/NewTwitchApi/Resources/UsersApi.php
@@ -75,21 +75,4 @@ class UsersApi extends AbstractResource
 
         return $this->callApi('users/follows', $queryParamsMap, $bearer);
     }
-
-    /**
-     * @throws GuzzleException
-     * @link https://dev.twitch.tv/docs/api/reference/#get-broadcaster-s-subscribers
-     */
-    public function getBroadcasterSubscriptions(string $broadcasterId, array $ids = [], string $bearer): ResponseInterface
-    {
-        $queryParamsMap = [];
-
-        $queryParamsMap[] = ['key' => 'broadcaster_id', 'value' => $broadcasterId];
-
-        foreach ($ids as $id) {
-            $queryParamsMap[] = ['key' => 'user_id', 'value' => $id];
-        }
-
-        return $this->callApi('subscriptions', $queryParamsMap, $bearer);
-    }
 }
